requirements.txt
Django==4.1.3
djangorestframework==3.14.0
psycopg2-binary==2.9.3
drf-spectacular==0.24.2